firebase_functions/manim-render/requirements.txt
typing_extensions
torchaudio
numpy==2.0.2
langsmith
manim==0.19.0
pybind11>=2.12
ManimPango==0.6.0
manim-voiceover==0.3.7
google-cloud-storage==2.5.0
requests==2.31.0
flask==2.2.5
openai==1.61.1